backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==10.0.1
pydantic==2.4.2
python-dotenv==1.0.0
jinja2>=3.1.0


# PostgreSQL & Database
psycopg2-binary==2.9.7
sqlalchemy==2.0.23
alembic==1.12.1
shortuuid==1.0.11
